ENH: added UseIndexSpace option to LabelmapSeeding module
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=14139
git-svn-id: http://svn.slicer.org/Slicer4/trunk@14139 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/vtkTeem/vtkSeedTracts.cxx
@@ -469,9 +469,9 @@ void vtkSeedTracts::SeedStreamlineFromPoint(double x,
 //----------------------------------------------------------------------------
 void vtkSeedTracts::SeedStreamlinesInROI()
 {
-  float idxX, idxY, idxZ;
-  float maxX, maxY, maxZ;
-  float gridIncX, gridIncY, gridIncZ;
+  double idxX, idxY, idxZ;
+  double maxX, maxY, maxZ;
+  double gridIncX, gridIncY, gridIncZ;
   int inExt[6];
   double point[3], point2[3];
